Move lib/node classes into lib/cloudtrax, and remove lib/node.py.
There wasn't much sense in having these classes under two separate
files, since the only part of the code that will be using Networks,
Nodes, etc. will be the CloudTrax instance itself.
